requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.6
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
python-dotenv~=1.0.0
django-filter~=23.3
djangorestframework-simplejwt
